feat(Snippet): controller interface
