--- src/main/kotlin/com/ingsis/snippets/config/AppConfig.kt
@@ -1,11 +1,0 @@
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.client.RestTemplate
-
-@Configuration
-open class AppConfig {
-  @Bean
-  open fun restTemplate(): RestTemplate {
-    return RestTemplate()
-  }
-}
